Refactor Edit organization components structure
