+++ apps/console/src/features/organizations/components/edit-organization/organization-list.tsx
@@ -49,10 +49,11 @@ import {
     history
 } from "../../core";
 import { getEmptyPlaceholderIllustrations } from "../../core/configs/ui";
-import { deleteOrganization } from "../api";
-import { OrganizationIcon } from "../configs";
-import { OrganizationManagementConstants } from "../constants";
-import { OrganizationInterface, OrganizationListInterface } from "../models";
+} from "../../../core";
+import { deleteOrganization } from "../../api";
+import { OrganizationIcon } from "../../configs";
+import { OrganizationManagementConstants } from "../../constants";
+import { OrganizationInterface, OrganizationListInterface } from "../../models";
 
 /**
  *
