+++ apps/console/src/features/organizations/components/index.ts
@@ -16,6 +16,6 @@
 * under the License.
 */
 
-export * from "./organization-list";
+export * from "./edit-organization/organization-list";
 export * from "./add-organization-modal";
 export * from "./organization-role-list";
