+++ apps/console/src/features/organizations/pages/organization-edit.tsx
@@ -27,7 +27,7 @@ import { useDispatch } from "react-redux";
 import { RouteChildrenProps } from "react-router-dom";
 import { AppConstants, FeatureConfigInterface, history } from "../../core";
 import { getOrganization } from "../api";
-import { EditOrganization } from "../components/edit-organization";
+import { EditOrganization } from "../components/edit-organization/edit-organization";
 import { OrganizationIcon } from "../configs";
 import { OrganizationManagementConstants } from "../constants";
 import { OrganizationResponseInterface } from "../models";
